// app/modules/shared/components/status-indicator/enums/status-indicator-values.enum.ts
export enum StatusIndicatorValues {
  VALID = 'Valid',
  INVALID = 'Invalid',
  UNCONFIGURED = 'Unconfigured',
  UNKNOWN = 'Unknown',
  CONNECTED = 'Connected',
  OFFLINE = 'Offline',
  CONNECTING = 'Connecting',
  ERROR = 'Error',
  DOWN = 'Down',
  GOOD = 'Good',
  WARNING = 'Warning',
  CRITICAL = 'Critical',
  POLLING_DISABLED = 'Polling Disabled',
  NA = 'N/A'
}